AI_MODELS/requirements.txt
flask==2.3.0
flask-cors==4.0.0
torch==2.0.0
torchvision==0.15.0
pillow==10.0.0
timm==0.9.0
opencv-python==4.8.0
numpy==1.24.0
